requirements.txt
flask==3.0.3
flask-cors==4.0.0
flask-sqlalchemy==3.1.1
sqlalchemy==2.0.30
python-dotenv==1.0.1
# note: no postgres driver by default; sqlite works out of the box.
# if you need postgres, pip install -r requirements-postgres.txt
